requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-dev -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anthropic==0.97.0
    # via runner-dashboard
anyio==4.13.0
    # via
    #   anthropic
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.7
    # via requests
click==8.3.3
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
distro==1.9.0
    # via anthropic
docstring-parser==0.18.0
    # via anthropic
fastapi==0.136.1
    # via runner-dashboard
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   anthropic
    #   runner-dashboard
idna==3.13
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via runner-dashboard
jiter==0.14.0
    # via anthropic
psutil==7.2.2
    # via runner-dashboard
pydantic==2.13.3
    # via
    #   anthropic
    #   fastapi
    #   runner-dashboard
pydantic-core==2.46.3
    # via pydantic
python-dotenv==1.2.2
    # via uvicorn
pyyaml==6.0.3
    # via
    #   runner-dashboard
    #   uvicorn
requests==2.33.1
    # via runner-dashboard
sniffio==1.3.1
    # via anthropic
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.46.0
    # via runner-dashboard
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn

prometheus-client==0.21.1
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
respx==0.21.1
